add independent image dataset
