+++ config/experiment/clustering.yaml
@@ -1,7 +1,7 @@
 # @package _global_
 defaults:
   - override /architecture: clustering
-  - override /dataset: v1_full
+  - override /dataset: v1_single
 
 loader:
-  batch_size: 1
+  batch_size: 8
